test: 修复第七轮测试审查发现的缺陷并补充覆盖 (#287-#296)
修复 10 个测试审查发现的缺陷：

#287 ScriptExecutor 路径穿越防护：添加 isValidInstanceId() 校验
#288 SqlExecutor 路径穿越防护 + libpq 连接串转义
#289 User::toJson() 移除 password_hash 字段，避免误用泄露
#290 Task/Workflow fromRow() 添加 NULL 字段处理，防止 pqxx 崩溃
#291 CronParser 步长循环用 long long 防止整数溢出无限循环
#292 ResourceCollector 添加 MemFree 回退（内核 < 3.14 无 MemAvailable）
#293 DagEngine findReadyTasks/findBlockedTasks 校验边端点存在性
#294 command/script/sql executor setpgid 返回值检查，防止 kill(-pid) 误杀 Worker
#295 CMakeLists.txt 注册 test_regression_116.cpp
#296 test_executor_extended.cpp 修复 8 处弱断言 + REQUIRE_NOTHROW lambda 模式

测试结果：504 个测试用例全部通过（2510 个断言）

+++ backend/src/scheduler/engine/cron_parser.h
@@ -289,8 +289,10 @@ private:
             }
 
             // Generate values
-            for (int v = range_start; v <= range_end; v += step) {
-                result.push_back(v);
+            // Fix #291: 用 long long 防止 v += step 整数溢出导致无限循环
+            // 例如 step=2147483647 时，int 的 v += step 会溢出为负数，导致 v <= range_end 恒真
+            for (long long v = range_start; v <= range_end; v += step) {
+                result.push_back(static_cast<int>(v));
             }
         }
